+++ tests/silence_lint_error/ruff_test.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 
+import sys
 from pathlib import Path
 
 import pytest
+from pytest_subprocess import FakeProcess
 
 from silence_lint_error.silence_lint_error import main
 
@@ -58,4 +60,23 @@ def foo():
     assert captured.err == """\
 -> finding errors with ruff
 no errors found
+"""
+
+
+def test_not_installed(capsys: pytest.CaptureFixture[str]):
+    with FakeProcess() as process:
+        process.register(
+            (sys.executable, '-mruff', process.any()),
+            returncode=1, stderr='/path/to/python3: No module named ruff\n',
+        )
+
+        ret = main(('ruff', 'F401', 'path/to/file.py'))
+
+    assert ret == 1
+
+    captured = capsys.readouterr()
+    assert captured.out == ''
+    assert captured.err == """\
+-> finding errors with ruff
+ERROR: /path/to/python3: No module named ruff
 """