+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- Apply auto-fixes with `ruff`.
+
 ## 1.2.1 (2023-12-11)
 
 ### Fixed