+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- Silence errors with `semgrep`.
+
 ## 1.3.0 (2023-12-15)
 
 ### Added