+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 1.4.1 (2024-05-09)
+
+The tests are no longer included in the built package.
+
 ## 1.4.0 (2024-04-11)
 
 ### Added